+++ CHANGES.rst
@@ -5,6 +5,8 @@
 6.0.1 (unreleased)
 ==================
 
+ - repozo: prevent an incorrect "option ignored" warning when running backup or verify.
+
 
 6.0 (2024-03-20)
 ================