+++ CHANGES.rst
@@ -5,6 +5,8 @@
 6.0.1 (unreleased)
 ==================
 
+ - repozo: fix restoring on stdout.
+
  - repozo: prevent an incorrect "option ignored" warning when running backup or verify.